requirements.txt
plaid-python>=14.0.0
anthropic>=0.25.0
gspread>=6.0.0
google-auth>=2.28.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
google-api-python-client>=2.120.0
requests>=2.31.0
